test(promql): refactor promqltest framework and add upstream test coverage
Refactored the promqltest framework into focused modules and added comprehensive test coverage from Prometheus upstream test suite.

Changes:
-Made dsl module and internal functions private. Only runner::run_test() remains public for generated tests.
- Refactored runner.rs by extracting loader, evaluator, and assertion logic into separate modules (loader.rs, evaluator.rs, assert.rs)
- Updated README.md with improved architecture documentation showing module responsibilities and execution flow
- Updated at_modifier.test from Prometheus upstream with ignore/resume blocks for unimplemented features
- Updated selectors.test from Prometheus upstream with ignore/resume blocks for unimplemented features

Test coverage:
- at_modifier.test: Tests @ modifier syntax with various scenarios
- Wrapped unimplemented: unary operator, subqueries, timestamp(),minute(), label_replace(), and aggregate functions
- selectors.test: Tests label selectors and matchers
- Wrapped unimplemented: full regex patterns (=~, !~), != with offset,binary operations, or operator, single-letter label names
